feat:ticket-reservation기능 DDD로 폴더구조 변경
