+++ src/main/java/com/flab/tiple/ticket/reservation/application/dto/request/TicketReservationRequestDto.java
@@ -1,0 +1,19 @@
+package com.flab.tiple.ticket.reservation.application.dto.request;
+
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class TicketReservationRequestDto {
+	private Long seatId;
+	private Long concertId;
+
+	@Builder
+	public TicketReservationRequestDto(Long seatId, Long concertId) {
+		this.seatId = seatId;
+		this.concertId = concertId;
+	}
+}
