+++ src/main/java/com/flab/tiple/ticket/reservation/domain/exception/TicketReservationStatusException.java
@@ -1,0 +1,10 @@
+package com.flab.tiple.ticket.reservation.domain.exception;
+
+import com.flab.tiple.global.exception.BusinessException;
+import com.flab.tiple.global.exception.ErrorCode;
+
+public class TicketReservationStatusException  extends BusinessException {
+	public TicketReservationStatusException(ErrorCode errorCode, String message) {
+		super(errorCode, message);
+	}
+}
