+++ src/main/java/com/flab/tiple/ticket/reservation/domain/model/enums/TicketProcessStatus.java
@@ -1,0 +1,6 @@
+package com.flab.tiple.ticket.reservation.domain.model.enums;
+
+public enum TicketProcessStatus {
+	SUCCESS,   // 예약 성공
+	WAITING,   // 웨이팅 등록
+}
