--- src/main/java/com/flab/tiple/ticket/reservation/enums/TicketProcessStatus.java
@@ -1,6 +1,0 @@
-package com.flab.tiple.ticket.reservation.enums;
-
-public enum TicketProcessStatus {
-	SUCCESS,   // 예약 성공
-	WAITING,   // 웨이팅 등록
-}
