--- src/main/java/com/flab/tiple/ticket/reservation/enums/TicketReservationStatus.java
@@ -1,8 +1,0 @@
-package com.flab.tiple.ticket.reservation.enums;
-
-public enum TicketReservationStatus {
-	PENDING,   // 좌석 예약 대기 상태
-	APPROVED,  // 예매 확정 상태
-	TIMEOUT,   // 예매 시간 만료 상태
-	CANCELLED  // 예매 취소 상태
-}
