+++ src/test/java/com/flab/tiple/global/config/RedissonConfig.java
@@ -23,7 +23,7 @@ import org.springframework.data.redis.listener.RedisMessageListenerContainer;
 
 import com.flab.tiple.global.listner.RedisKeyExpirationListener;
 import com.flab.tiple.global.util.RedissionLockExecutor;
-import com.flab.tiple.ticket.reservation.repository.TicketReservationRedisRepository;
+import com.flab.tiple.ticket.reservation.domain.repository.TicketReservationRedisRepository;
 import com.flab.tiple.ticket.waiting.repository.TicketWaitingRedisRepository;
 
 import lombok.extern.slf4j.Slf4j;
