+++ src/test/java/com/flab/tiple/ticket/controller/TicketReservationControllerTest.java
@@ -41,24 +41,21 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.flab.tiple.concert.dto.response.ConcertSeatInfo;
 import com.flab.tiple.concert.enums.ConcertSeatGrade;
-import com.flab.tiple.concert.exception.ConcertRemainSeatExistException;
 import com.flab.tiple.global.auth.aop.LoginCheckAspect;
 import com.flab.tiple.global.config.MockBeanConfig;
 import com.flab.tiple.global.config.TestSecurityConfig;
-import com.flab.tiple.global.exception.ErrorCode;
 import com.flab.tiple.global.response.ApiResponse;
 import com.flab.tiple.global.security.JwtTokenProvider;
 import com.flab.tiple.global.security.SecurityConfig;
 import com.flab.tiple.global.util.JwtTokenUtil;
 import com.flab.tiple.member.dto.response.MemberInfoDto;
-import com.flab.tiple.ticket.reservation.controller.TicketReservationController;
-import com.flab.tiple.ticket.reservation.dto.request.TicketReservationRequestDto;
-import com.flab.tiple.ticket.reservation.dto.response.TicketReservationInfoResponseDto;
-import com.flab.tiple.ticket.reservation.dto.response.TicketReservationResponseDto;
-import com.flab.tiple.ticket.reservation.enums.TicketProcessStatus;
-import com.flab.tiple.ticket.reservation.enums.TicketReservationStatus;
-import com.flab.tiple.ticket.reservation.facade.TicketReservationFacade;
-import com.flab.tiple.ticket.reservation.service.TicketReservationService;
+import com.flab.tiple.ticket.reservation.presentation.api.TicketReservationController;
+import com.flab.tiple.ticket.reservation.application.dto.request.TicketReservationRequestDto;
+import com.flab.tiple.ticket.reservation.application.dto.response.TicketReservationInfoResponseDto;
+import com.flab.tiple.ticket.reservation.application.dto.response.TicketReservationResponseDto;
+import com.flab.tiple.ticket.reservation.domain.model.enums.TicketProcessStatus;
+import com.flab.tiple.ticket.reservation.domain.model.enums.TicketReservationStatus;
+import com.flab.tiple.ticket.reservation.application.facade.TicketReservationFacade;
 import com.flab.tiple.ticket.waiting.dto.response.TicketWaitingResponseDto;
 
 @WebMvcTest(
