+++ src/test/java/com/flab/tiple/ticket/service/TicketReservationServiceTest.java
@@ -33,17 +33,17 @@ import com.flab.tiple.member.domain.Member;
 import com.flab.tiple.member.repository.MemberRepository;
 import com.flab.tiple.sse.dto.NotificationDto;
 import com.flab.tiple.sse.service.SseEmitterService;
-import com.flab.tiple.ticket.reservation.domain.TicketReservation;
-import com.flab.tiple.ticket.reservation.domain.TicketReservationRedis;
-import com.flab.tiple.ticket.reservation.dto.request.TicketReservationRequestDto;
-import com.flab.tiple.ticket.reservation.dto.response.TicketReservationInfoResponseDto;
-import com.flab.tiple.ticket.reservation.dto.response.TicketReservationResponseDto;
-import com.flab.tiple.ticket.reservation.enums.TicketProcessStatus;
-import com.flab.tiple.ticket.reservation.enums.TicketReservationStatus;
-import com.flab.tiple.ticket.reservation.exception.TicketReservationNotFoundException;
-import com.flab.tiple.ticket.reservation.repository.TicketReservationRedisRepository;
-import com.flab.tiple.ticket.reservation.repository.TicketReservationRepository;
-import com.flab.tiple.ticket.reservation.service.TicketReservationServiceImpl;
+import com.flab.tiple.ticket.reservation.domain.model.TicketReservation;
+import com.flab.tiple.ticket.reservation.domain.model.TicketReservationRedis;
+import com.flab.tiple.ticket.reservation.application.dto.request.TicketReservationRequestDto;
+import com.flab.tiple.ticket.reservation.application.dto.response.TicketReservationInfoResponseDto;
+import com.flab.tiple.ticket.reservation.application.dto.response.TicketReservationResponseDto;
+import com.flab.tiple.ticket.reservation.domain.model.enums.TicketProcessStatus;
+import com.flab.tiple.ticket.reservation.domain.model.enums.TicketReservationStatus;
+import com.flab.tiple.ticket.reservation.domain.exception.TicketReservationNotFoundException;
+import com.flab.tiple.ticket.reservation.domain.repository.TicketReservationRedisRepository;
+import com.flab.tiple.ticket.reservation.domain.repository.TicketReservationRepository;
+import com.flab.tiple.ticket.reservation.domain.service.TicketReservationServiceImpl;
 import com.flab.tiple.ticket.waiting.domain.TicketWaitingRedis;
 import com.flab.tiple.ticket.waiting.dto.response.TicketWaitingResponseDto;
 import com.flab.tiple.ticket.waiting.enums.TicketWaitingStatus;
